+++ cloudapp/src/app/user-controls/search-form/search-form.component.scss
@@ -31,7 +31,8 @@
 
 .form-card-field-small {
   width: 17%;
-  flex-grow: 2;
+  // max-width: 130px;
+  flex-grow: 1;
   min-width: 100px;
 }
